+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 * FEATURE: advanced work filtering (publication, commencement, repeal, amendment etc.)
 * FEATURE: show offline warning when editing a document
 * FEATURE: site sidebar removed and replaced with tabs
+* FEATURE: show source attachments and work publication document side-by-side when editing a document
 * New schedule syntax makes headings and subheadings clearer
 * Move document templates from templates/documents/ to templates/indigo_api/documents/